# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: GTW IOT API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
securityDefinitions:
  UserSecurity:
    type: apiKey
    in: header
    name: x-api-key
security:
  - UserSecurity: []
paths:
...
  /device/getDevice:
    x-swagger-router-controller: deviceManage
    get:
      tags:
      - device
      summary: get all device
      description: get all device by AssetID
      # used as the method name of the controller
      operationId: getDevice
      parameters:
      - name: AssetID
        in: query
        description: User object thats need to be deleted
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/DeviceModel"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /device/addDevice:
    x-swagger-router-controller: deviceManage
    post:
      tags:
      - device
      summary: create a new device
      description: create a new device
      # used as the method name of the controller
      operationId: addDevice
      parameters:
      - in: body
        name: body
        description: Device object thats need to be added
        required: true
        schema:
          $ref: '#/definitions/DeviceModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /device/updateDevice:
    x-swagger-router-controller: deviceManage
    put:
      tags:
      - device
      summary: update a Device
      description: update a Device
      # used as the method name of the controller
      operationId: updateDevice
      parameters:
      - in: body
        name: body
        description: User object thats need to be updated
        required: true
        schema:
          $ref: '#/definitions/DeviceModel'
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /device/deleteDevice:
    x-swagger-router-controller: deviceManage
    delete:
      tags:
      - device
      summary: delete a Device
      description: delete a Device
      # used as the method name of the controller
      operationId: deleteDevice
      parameters:
      - name: DeviceID
        in: query
        description: Device object thats need to be deleted
        required: true
        type: string
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  DeviceModel:
    type: object
    properties:
      DeviceID:
        type: string
      Type:
        type: string
      AssetID:
        type: string
      Location:
        type: string
      Location_Display:
        type: string
      LinkDeviceID:
        type: string
      Tag:
        type: string
...
  SuccessResponse:
    required:
      - message
    properties:
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string